+++ src/main/java/com/verdemar/exception/apartment/ApartmentNotFoundException.java
@@ -2,7 +2,7 @@ package com.verdemar.exception.apartment;
 
 public class ApartmentNotFoundException extends RuntimeException {
 
-  public ApartmentNotFoundException(Short id) {
+  public ApartmentNotFoundException(Integer id) {
     super("Apartment not found with id: " + id);
   }
