+++ src/main/java/com/verdemar/repository/BookingRepository.java
@@ -12,7 +12,7 @@ public interface BookingRepository extends JpaRepository<Booking, Long> {
   // Encuentra todas las fechas de reserva para un apartmento específico
   @Query(
       "SELECT new com.verdemar.domain.dto.BookingDateRange(b.startDate, b.endDate) FROM Booking b WHERE b.apartment.id = :apartmentId")
-  List<BookingDateRange> findAllDatesByApartment(@Param("apartmentId") Short apartmentId);
+  List<BookingDateRange> findAllDatesByApartment(@Param("apartmentId") Integer apartmentId);
   
   // Encuentra reservas que se solapan con un rango de fechas dado
   @Query("SELECT b FROM Booking b WHERE b.startDate < :endDate AND b.endDate > :startDate")
@@ -30,7 +30,7 @@ public interface BookingRepository extends JpaRepository<Booking, Long> {
       ORDER BY b.startDate
       """)
   List<Booking> findBookingsByApartmentAndMonth(
-      @Param("apartmentId") Short apartmentId,
+      @Param("apartmentId") Integer apartmentId,
       @Param("month") int month,
       @Param("year") int year);
