+++ src/main/java/com/verdemar/repository/PriceRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PriceRepository extends JpaRepository<Price, PriceId> {
     List<Price> findByApartmentIdAndDateBetween(
-    Short apartmentId,
+    Integer apartmentId,
     LocalDate startDate,
     LocalDate endDate
 );
